refactor: extract streaming-service categorization into recommenders/streaming.py (2.10.44)
PR2 step 3 of the external.py architecture decomposition - pure relocation of categorize_by_streaming_service. streaming.py imports get_watch_providers from huntarr.py. external.py still calls it directly from process_user/process_user_movie_library/process_user_tv_library. Verified byte-identical against the golden-output harness.
